# platform-backend/docs/compute-openapi.yaml
openapi: 3.1.0
info:
  title: Fractal Compute private v1 API
  version: 1.0.0
  description: |
    Authoritative Platform Worker contract for the current C++ Compute v1 server.
    Legacy /api/* routes are explicitly outside this contract.
servers:
  - url: http://compute.internal
security:
  - ComputeServiceKey: []
paths:
...
  /compute/v1/capabilities:
    get:
      operationId: capabilities
      responses:
        '200':
          description: Authenticated renderer capabilities
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    ComputeServiceKey:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-key
...
  schemas:
...
    Problem:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
  responses:
    Problem:
      description: Stable Compute error envelope
      content:
        application/json: { schema: { $ref: '#/components/schemas/Problem' } }